+++ R/utils.R
@@ -502,14 +502,15 @@ getTimeRange <- function(x, mode=c('event', 'detection'), sample=FALSE) {
                    nrow(d) == 0) {
                     return(NULL)
                 }
-                out <- d[, c('UID', 'UTC', 'duration'), drop = FALSE]
+                out <- d[, c('UID', 'UTC'), drop = FALSE]
                 if('duration' %in% colnames(d)) {
                     switch(attr(d, 'calltype'),
                            'whistle' = out$duration <- d$duration,
                            # 'click' = out$duration <- d$duration / 1e6,
                            'click' = out$duration <- 0, # duration is not reliable for clicks
                            'cepstrum' = out$duration <- d$duration,
-                           'gpl' = out$duration <- d$duration
+                           'gpl' = out$duration <- d$duration,
+                           out$duration <- 0
                     )
                 } else {
                     out$duration <- 0